Noseyparker (#2)
* add support for noseyparker
* add arguments to skip specific tools, another arg to open HTML report in default browser when done
* lots of refactoring, rename tool to secretsynth
* fixing merg logic for noseyparker results. style the html report a teensy bit
* cleaning up aggregated metrics and report summary
* fixing edge cases and error handling.
* added an experimental file for future reference. trying to do some GHAS pattern matching on secrets found in other tools

+++ org-scan/ghas_secret_alerts_fetch.py
@@ -26,7 +26,15 @@ def fetch_repos(account_type, account, headers, logger=None, page=1, per_page=10
     
     return repos
 
-def fetch_ghas_secret_scanning_alerts(owner_type, owners, headers, report_name, logger=None):
+def fetch_ghas_secret_scanning_alerts(owner_type, 
+                                      owners, headers, 
+                                      report_name, 
+                                      dry_run=False, 
+                                      logger=None):
+    
+    if dry_run:
+        print(f"dry-run: Calling Github REST API for all repos under orgs: {owners}")
+        return
     
     # Open the CSV file
     with open(report_name, 'w', newline='') as csvfile:
